login: give session ownership of VT
The tty associated with a VT should be owned by the owner of the session
running on the VT. This is important for supporting a socket activated X
server, since the X server will open the tty itself.

This commit makes sure to chown the tty any time a session is
created and and chown it back to root any time the session
is removed. This behavior is copied from /bin/login.

(cherry picked from commit d6176c6c97bf0614c2e7caaf2156bf813b39337a)