systemd-nspawn: Optionally exit on container reboot
In NixOS container units, we need to run some host-side network setup
after the container has been started, specifically to initialise the
host side of virtual Ethernet pairs. If the container is rebooted, we
need to re-run this initialisation. Thus it's better if systemd-nspawn
just exits so that the container unit can be restarted in its
entirety.